requirements.txt
influxdb-client==1.30.0
python-dotenv==1.0.0
